--- a/src/core/builtins/builtins_ffmpeg_bitstream_filters.ml
+++ b/src/core/builtins/builtins_ffmpeg_bitstream_filters.ml
@@ -84,7 +84,7 @@ let process (type a) ~put_data ~(mk_params : a mk_params)
       packets
   in
   let data =
-    { Ffmpeg_content_base.params = Some (mk_params params); data; length }
+    { Content_video.Base.params = Some (mk_params params); data; length }
   in
   let data = Ffmpeg_copy_content.lift_data data in
   put_data generator data
@@ -121,7 +121,7 @@ let on_data (type a)
        a handler) ~put_data ~(get_params : a get_params)
     ~(get_packet : a get_packet) ~(mk_params : a mk_params)
     ~(mk_packet : a mk_packet) ~generator
-    ({ Ffmpeg_content_base.params; data } : Ffmpeg_copy_content.data) =
+    ({ Content_video.Base.params; data } : Ffmpeg_copy_content.data) =
   List.iter
     (fun (_, { Ffmpeg_copy_content.stream_idx; time_base; packet }) ->
       let handler =