--- a/src/core/decoder/gstreamer_decoder.ml
+++ b/src/core/decoder/gstreamer_decoder.ml
@@ -138,9 +138,8 @@ let create_decoder ?(merge_tracks = false) _ ~width ~height ~channels ~mode
       let y_stride = round4 width in
       let uv_stride = round4 (width / 2) in
       let img = Image.YUV420.make_data width height buf y_stride uv_stride in
-      let stream = Video.Canvas.single_image img in
       let fps = { Decoder.num = Lazy.force Frame.video_rate; den = 1 } in
-      buffer.Decoder.put_yuva420p ~fps stream);
+      buffer.Decoder.put_yuva420p ~fps (Video.Canvas.Image.make img));
     GU.flush ~log gst.bin
   in
   let seek off =