## Summary

This PR makes it possible to define sub-video sample frames.

Typically, this makes it possible to have a latency of `0.01s` with audio/video scripts.

The trade-off in CPU vs. memory usage is not as great as expected but, nonetheless, I believe that is fundamental that we should be able to do low-latency stream generation.

Video memory, on the other hand, is greatly improved. Not because if the change of frame size but because of the more efficient APIs to generate content without allocating empty frames. See below for the report.

### TODO

- [x] Cleanup, proof read
- [x] Large video example.

## API changes

In order to make this work, these changes:

### Merge the video payload from ffmpeg into plain video content

This means that we are not longer using `Video.Canvas` but our own list of `(position, image)`. This is pretty much the same but allows us to keep track of the position of an image within the content for easy content sub/append.

### Video frame normalization

Once we start having frames that are sub-video sample, it is possible to have:
* Frames without video image, for instance:
```
frame size = 0.01s = 441 samples
video rate = 25 = 1 video image each 4 frames

|-------|------|------|------|
| i     |      |      |      |
|-------|------|------|------|
```

In dynamic swiching context, it is then possible to append two frames with initial image from two different sources, leading to situations like:

```
|-------|------|------|------|
| i  i  |      |      |   i  |
|-------|------|------|------|
```

Things become even more complicated when dealing with operators such as `add`.

Thus, in order to remediate this, a new `video generator` facility is created. This API keeps track of its current position and is able to generate video with the expected frame rate. It also takes a `create` function which is used to return the image that should be used when creating video content.

Using this new API we can normalize all source's video content as follow:
* All sources keep track of the last image they have produced.
* When producing a video frame, the source calls its own `video generator` with that frame's length and uses the closest image from the generated content, if available, or its latest image, to generate the new content.

This, essentially, means that all sources are video-rate controlled using a nearest sample mechanism.

This makes it possible to normalize assumptions such as having images available to consistently add them in the `video.add` operators.

Here are the new API:
```ocaml
       method private generate_video :
         ?create:
           (pos:int -> width:int -> height:int -> unit -> Video.Canvas.image) ->
         field:Frame.Fields.field ->
         int ->
         Content.Video.data

       method last_image : Frame.Fields.field -> Video.Canvas.image

       method private nearest_image :
         pos:int ->
         last_image:Video.Canvas.image ->
         Content.Video.data ->
         Video.Canvas.image
```

And here's how we normalize each frame while generating content:

```ocaml
    method private normalize_video ~field content =
      let buf = Content.Video.get_data content in
      let data = buf.Content.Video.data in
      let last_image =
        match List.rev data with
          | (_, img) :: _ ->
              self#set_last_image ~field img;
              img
          | [] -> self#last_image field
      in
      Content.Video.lift_data
        (self#internal_generate_video ~field ~priv:true
           ~create:(fun ~pos ~width:_ ~height:_ () ->
             self#nearest_image ~pos ~last_image buf)
           (Content.length content))

    method private normalize_video_content =
      Frame.Fields.mapi (fun field content ->
          if Content.Video.is_data content then
            self#normalize_video ~field content
          else content)
```

### Mapping for content manipulation

On the consumer side, manipulating the content becomes a simple matter of mapping the list of images:
```ocaml
    method private generate_frame =
      let buf =
        Content.Video.get_data (source#get_mutable_content Frame.Fields.video)
      in
      params ();
      source#set_frame_data Frame.Fields.video Content.Video.lift_data
        {
          buf with
          Content.Video.data =
            List.map
              (fun (pos, img) -> (pos, self#render img))
              buf.Content.Video.data;
        }
```

## Performances

First tests seem to show that performance is the same for a `0.04s` frame and decreases memory usage by about `10%` with a `0.01s` frame at the cost of an increased CPU usage.

Script:
```liquidsoap
def f() =
  s = playlist("~/sources/test-stream/audio")
  s = crossfade(s)
  output.dummy(fallible=true, s)
end

for i = 0 to 100 do
  thread.run(delay=float_of_int(i), f)
end

output.dummy(blank())
```

### `main`, `0.04s` frame

<img width="229" alt="Screenshot 2023-12-31 at 9 38 15 AM" src="https://github.com/savonet/liquidsoap/assets/871060/399e052b-6b0c-4ad4-bc29-9d896a59a85f">

<img width="1288" alt="Screenshot 2023-12-31 at 9 38 27 AM" src="https://github.com/savonet/liquidsoap/assets/871060/b6e99389-4fd4-4cd7-96dd-ad0f39cf6a1b">

### This PR, `0.04s` frame:

<img width="314" alt="Screenshot 2023-12-31 at 9 41 32 AM" src="https://github.com/savonet/liquidsoap/assets/871060/59569da8-ea8f-43fa-9a14-022294179f3a">

<img width="1286" alt="Screenshot 2023-12-31 at 9 41 45 AM" src="https://github.com/savonet/liquidsoap/assets/871060/c058e4e7-6402-4af0-9210-4b61eeca6b70">

### This PR, `0.01s` frame

<img width="301" alt="Screenshot 2023-12-31 at 9 34 38 AM" src="https://github.com/savonet/liquidsoap/assets/871060/4a240e72-534b-4967-841b-a7220e0b40d1">

<img width="1274" alt="Screenshot 2023-12-31 at 9 34 53 AM" src="https://github.com/savonet/liquidsoap/assets/871060/1f9f9f46-5228-48ce-bb23-f78e34e9b6ef">

## Complex video performace

This is the script that generates this:

<img width="1281" alt="Screenshot 2023-12-31 at 5 12 14 PM" src="https://github.com/savonet/liquidsoap/assets/871060/38c54ac9-7a87-4ccb-b4f7-79b21635ebda">

### Main, `0.04s` frame

<img width="321" alt="Screenshot 2023-12-31 at 5 09 58 PM" src="https://github.com/savonet/liquidsoap/assets/871060/a7be4fbf-1a4d-4c62-8d48-b71894e993e9">

<img width="815" alt="Screenshot 2023-12-31 at 5 10 08 PM" src="https://github.com/savonet/liquidsoap/assets/871060/12953337-2cb6-4688-b7de-8f3ca324509f">


### This PR, `0.04s` frame

<img width="325" alt="Screenshot 2023-12-31 at 5 11 37 PM" src="https://github.com/savonet/liquidsoap/assets/871060/e22254f2-605f-48a2-99ce-8662fc4db6c1">

<img width="1133" alt="Screenshot 2023-12-31 at 5 11 50 PM" src="https://github.com/savonet/liquidsoap/assets/871060/30227140-7409-451b-8ac7-fbb609264288">